+++ plugins/aws/src/index.ts
@@ -1,2 +1,2 @@
 export * from "./read";
-export * from "./upload";
+export * from "./uploadS3";
